dom/tests/mochitest/geolocation/test_timeoutWatch.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=455327
-->
<head>
  <title>Test for timeout option </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=455327">Mozilla Bug 478911</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug  **/

// ensure we are using the right testing provider
resume_geolocationProvider();
force_prompt(true);

SimpleTest.waitForExplicitFinish();

stop_geolocationProvider();

function successCallback(pos){
  ok(false, "success should have never been called.");
  resume_geolocationProvider();
  reset_prompt()
  SimpleTest.finish();
}

function errorCallback(err) {
  if (err.code == err.POSITION_UNAVAILABLE)
    ok(false, "nothing is hooked up to test against.");
  else
    ok(err.code == err.TIMEOUT, "ensure error is a timeout.");
  resume_geolocationProvider();
  reset_prompt()
  SimpleTest.finish();
}


var options = {
    maximumAge: 0,
    timeout: 10
};

navigator.geolocation.watchPosition(successCallback,
                                    errorCallback,
                                    options);
</script>
</pre>
</body>
</html>
